requirements.txt
fastapi==0.95.0
numpy==1.24.2
opencv_python==4.7.0.72
Pillow==9.5.0
torch==2.0.0
uvicorn==0.21.1